Cleans up the parameters package for generic volumes.
