+++ pkg/parameters/parameters.go
@@ -27,164 +27,11 @@ import (
 	"sigs.k8s.io/gcp-compute-persistent-disk-csi-driver/pkg/convert"
 )
 
-const (
-	// Disk Params
-	ParameterAccessMode = "access-mode"
-
-	// Parameters for StorageClass
-	ParameterKeyType                          = "type"
-	ParameterKeyReplicationType               = "replication-type"
-	ParameterKeyDiskEncryptionKmsKey          = "disk-encryption-kms-key"
-	ParameterKeyLabels                        = "labels"
-	ParameterKeyProvisionedIOPSOnCreate       = "provisioned-iops-on-create"
-	ParameterKeyProvisionedThroughputOnCreate = "provisioned-throughput-on-create"
-	ParameterAvailabilityClass                = "availability-class"
-	ParameterKeyEnableConfidentialCompute     = "enable-confidential-storage"
-	ParameterKeyStoragePools                  = "storage-pools"
-	ParameterKeyUseAllowedDiskTopology        = "use-allowed-disk-topology"
-
-	// Parameters for Data Cache
-	ParameterKeyDataCacheSize               = "data-cache-size"
-	ParameterKeyDataCacheMode               = "data-cache-mode"
-	ParameterKeyResourceTags                = "resource-tags"
-	ParameterKeyEnableMultiZoneProvisioning = "enable-multi-zone-provisioning"
-
-	// Parameters for VolumeSnapshotClass
-	ParameterKeyStorageLocations = "storage-locations"
-	ParameterKeySnapshotType     = "snapshot-type"
-	ParameterKeyImageFamily      = "image-family"
-	replicationTypeNone          = "none"
-
-	// Keys for PV and PVC parameters as reported by external-provisioner
-	ParameterKeyPVCName      = "csi.storage.k8s.io/pvc/name"
-	ParameterKeyPVCNamespace = "csi.storage.k8s.io/pvc/namespace"
-	ParameterKeyPVName       = "csi.storage.k8s.io/pv/name"
-
-	// Keys for tags to put in the provisioned disk description
-	tagKeyCreatedForClaimNamespace = "kubernetes.io/created-for/pvc/namespace"
-	tagKeyCreatedForClaimName      = "kubernetes.io/created-for/pvc/name"
-	tagKeyCreatedForVolumeName     = "kubernetes.io/created-for/pv/name"
-	tagKeyCreatedBy                = "storage.gke.io/created-by"
-
-	// Keys for Snapshot and SnapshotContent parameters as reported by external-snapshotter
-	ParameterKeyVolumeSnapshotName        = "csi.storage.k8s.io/volumesnapshot/name"
-	ParameterKeyVolumeSnapshotNamespace   = "csi.storage.k8s.io/volumesnapshot/namespace"
-	ParameterKeyVolumeSnapshotContentName = "csi.storage.k8s.io/volumesnapshotcontent/name"
-
-	// Parameters for AvailabilityClass
-	ParameterNoAvailabilityClass       = "none"
-	ParameterRegionalHardFailoverClass = "regional-hard-failover"
-
-	// Keys for tags to put in the provisioned snapshot description
-	tagKeyCreatedForSnapshotName        = "kubernetes.io/created-for/volumesnapshot/name"
-	tagKeyCreatedForSnapshotNamespace   = "kubernetes.io/created-for/volumesnapshot/namespace"
-	tagKeyCreatedForSnapshotContentName = "kubernetes.io/created-for/volumesnapshotcontent/name"
-
-	// Hyperdisk disk types
-	DiskTypeHdHA = "hyperdisk-balanced-high-availability"
-	DiskTypeHdT  = "hyperdisk-throughput"
-	DiskTypeHdE  = "hyperdisk-extreme"
-	DiskTypeHdML = "hyperdisk-ml"
-
-	// Parameters for VolumeSnapshotClass
-	DiskSnapshotType = "snapshots"
-	DiskImageType    = "images"
-)
-
-type StoragePool struct {
-	Project      string
-	Zone         string
-	Name         string
-	ResourceName string
-}
-
-type DataCacheParameters struct {
-	// Values: {string} in int64 form
-	// Default: ""
-	DataCacheSize string
-	// Values: writethrough, writeback
-	// Default: writethrough
-	DataCacheMode string
-}
-
-// DiskParameters contains normalized and defaulted disk parameters
-type DiskParameters struct {
-	// Values: pd-standard, pd-balanced, pd-ssd, or any other PD disk type. Not validated.
-	// Default: pd-standard
-	DiskType string
-	// Values: "none", regional-pd
-	// Default: "none"
-	ReplicationType string
-	// Values: {string}
-	// Default: ""
-	DiskEncryptionKMSKey string
-	// Values: {map[string]string}
-	// Default: ""
-	Tags map[string]string
-	// Values: {map[string]string}
-	// Default: ""
-	Labels map[string]string
-	// Values: {int64}
-	// Default: none
-	ProvisionedIOPSOnCreate int64
-	// Values: {int64}
-	// Default: none
-	ProvisionedThroughputOnCreate int64
-	// Values: {bool}
-	// Default: false
-	EnableConfidentialCompute bool
-	// Default: false
-	ForceAttach bool
-	// Values: {[]string}
-	// Default: ""
-	StoragePools []StoragePool
-	// Values: {map[string]string}
-	// Default: ""
-	ResourceTags map[string]string
-	// Values: {bool}
-	// Default: false
-	MultiZoneProvisioning bool
-	// Values: READ_WRITE_SINGLE, READ_ONLY_MANY, READ_WRITE_MANY
-	// Default: READ_WRITE_SINGLE
-	AccessMode string
-	// Values {}
-	// Default: false
-	UseAllowedDiskTopology bool
-}
-
-func (dp *DiskParameters) IsRegional() bool {
-	return dp.ReplicationType == "regional-pd" || dp.DiskType == DiskTypeHdHA
-}
-
-// SnapshotParameters contains normalized and defaulted parameters for snapshots
-type SnapshotParameters struct {
-	StorageLocations []string
-	SnapshotType     string
-	ImageFamily      string
-	Tags             map[string]string
-	Labels           map[string]string
-	ResourceTags     map[string]string
-}
-
-type ParameterProcessor struct {
-	DriverName         string
-	EnableStoragePools bool
-	EnableMultiZone    bool
-	EnableHdHA         bool
-	EnableDiskTopology bool
-}
-
-type ModifyVolumeParameters struct {
-	IOPS       *int64
-	Throughput *int64
-}
-
 // ExtractAndDefaultParameters will take the relevant parameters from a map and
 // put them into a well defined struct making sure to default unspecified fields.
 // extraVolumeLabels are added as labels; if there are also labels specified in
 // parameters, any matching extraVolumeLabels will be overridden.
 func (pp *ParameterProcessor) ExtractAndDefaultParameters(parameters map[string]string, extraVolumeLabels map[string]string, enableDataCache bool, extraTags map[string]string) (DiskParameters, DataCacheParameters, error) {
-
 	p := DiskParameters{
 		DiskType:             "pd-standard",           // Default
 		ReplicationType:      replicationTypeNone,     // Default
@@ -275,7 +122,7 @@ func (pp *ParameterProcessor) ExtractAndDefaultParameters(parameters map[string]
 			if paramEnableConfidentialCompute {
 				// DiskEncryptionKmsKey is needed to enable confidentialStorage
 				if val, ok := parameters[ParameterKeyDiskEncryptionKmsKey]; !ok || !isValidDiskEncryptionKmsKey(val) {
-					return p, d, fmt.Errorf("Valid %v is required to enable ConfidentialStorage", ParameterKeyDiskEncryptionKmsKey)
+					return p, d, fmt.Errorf("valid %v is required to enable ConfidentialStorage", ParameterKeyDiskEncryptionKmsKey)
 				}
 			}
 
@@ -304,7 +151,7 @@ func (pp *ParameterProcessor) ExtractAndDefaultParameters(parameters map[string]
 			d.DataCacheSize = strconv.FormatInt(paramDataCacheSize, 10)
 		case ParameterKeyDataCacheMode:
 			if !enableDataCache {
-				return p, d, fmt.Errorf("data caching enabled %v; parameters contains invalid option %q", enableDataCache, ParameterKeyDataCacheSize)
+				return p, d, fmt.Errorf("data caching enabled %v; parameters contains invalid option %q", enableDataCache, ParameterKeyDataCacheMode)
 			}
 			if err := ValidateDataCacheMode(v); err != nil {
 				return p, d, fmt.Errorf("parameters contains invalid option: %s: %w", ParameterKeyDataCacheMode, err)
@@ -425,7 +272,6 @@ func extractResourceTagsParameter(tagsString string, resourceTags map[string]str
 }
 
 func ExtractModifyVolumeParameters(parameters map[string]string) (ModifyVolumeParameters, error) {
-
 	modifyVolumeParams := ModifyVolumeParameters{}
 
 	for key, value := range parameters {
@@ -446,6 +292,7 @@ func ExtractModifyVolumeParameters(parameters map[string]string) (ModifyVolumePa
 			return ModifyVolumeParameters{}, fmt.Errorf("parameters contain unknown parameter: %s", key)
 		}
 	}
+
 	return modifyVolumeParams, nil
 }
 
@@ -457,13 +304,13 @@ func convertStringToAvailabilityClass(str string) (string, error) {
 	case ParameterRegionalHardFailoverClass:
 		return ParameterRegionalHardFailoverClass, nil
 	}
-	return "", fmt.Errorf("Unexpected boolean string %s", str)
+	return "", fmt.Errorf("unexpected boolean string %s", str)
 }
 
 // StoragePoolZones returns the unique zones of the given storage pool resource names.
 // Returns an error if multiple storage pools in 1 zone are found.
 func StoragePoolZones(storagePools []StoragePool) ([]string, error) {
-	zonesSet := sets.String{}
+	zonesSet := sets.New[string]()
 	var zones []string
 	for _, sp := range storagePools {
 		if zonesSet.Has(sp.Zone) {
